% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation}
\usage{
translation(codons)
}
\arguments{
\item{codons}{codon input generated from codon_start}
}
\value{
protein returns protein, based on codon input
}
\description{
Translation
}
